API

Collector

EmeraldData.Collector.clean_database! — Function

clean_database!(selection::String = "old")
clean_database!(arttags::Vector{String})

Cleans the local database of datasets by removing old or specified datasets.

EmeraldData.Collector.download_dataset! — Function

Download the dataset file for a given artifact tag

EmeraldData.Collector.sync_database! — Function

Synchronize the local database with the remote database and download the datasets

Indexer

EmeraldData.Indexer.lat_ind — Function

lat_ind(lat::Number, res::Number)

Compute the latitude index from -90 to 90 degrees, given

lat: latitude in degrees

res: resolution in degrees (per grid)

EmeraldData.Indexer.lon_ind — Function

lon_ind(lon::Number, res::Number)

Compute the longitude index from -180 to 180 degrees, given

lon: longitude in degrees (if >180, it will be converted to -180 to 180)

res: resolution in degrees (per grid)

EmeraldData.Indexer.read_LUT — Function

readLUT() is an alias of readdataset()

Requestor

EmeraldData.Requestor.request_site_data — Function

request_site_data(server::String, user::String, tag::String, lat::Number, lon::Number, cycle::Int = 0)

Send a request to the specified server to get the artifact data for the given location, given

server the server URL (e.g., "http://localhost:8000")

user the user name

tag the artifact tag (e.g., "LM4X1Y_V1")

lat the target latitude

lon the target longitude

cycle the data cycle number (0 for all cycles, default is 0)